break all db related code out into db.js
